refactor(Map): Extract Search Marker logic to own hook
